requirements.txt

black==22.3.0
click==8.1.3
pytest==7.4.0
pytest-cov==4.0.0
pylint==2.15.3
streamlit
#tempfile
#time 
#rust based linter
ruff==0.0.284
python-dotenv
google.generativeai
numpy
opencv-python == 4.10.0.84
ultralytics == 8.3.27